Update dependencies, add new pages and components, and refactor code.
