+++ src/pages/Profile/Profile.jsx
@@ -1,4 +1,14 @@
-import React from 'react'
+// Autor: Esteban Maya
+
+import { Link, useParams } from "react-router-dom";
+import Gallery from "../../components/Gallery";
+import Card from "../../components/Card";
+import { useCallback, useContext, useEffect, useState } from 'react';
+import { AuthContext } from '../../context/AuthContext';
+
+
+
+
 
 const Profile = () => {
   return (
